refactor: move database to root of project.
