+++ lib/data/data.dart
@@ -1,15 +1,3 @@
-export 'datasources/local/database.dart';
-export 'models/novel.dart';
-export 'models/status.dart';
-export 'models/work_type.dart';
-export 'models/reading_direction.dart';
-export 'models/category.dart';
-export 'models/volume.dart';
-export 'models/chapter.dart';
-export 'models/meta_entry.dart';
-export 'models/namespaces.dart';
-export 'models/assets.dart';
-export 'models/updates.dart';
 export 'repositories/novel/gateway_repository_impl.dart';
 export 'repositories/novel/novel_repository_impl.dart';
 export 'repositories/network_repository_impl.dart';
