+++ lib/data/exception.dart
@@ -1,6 +1,1 @@
-/// This exception is thrown when the database seeding
-/// is not as expected.
-class SeedException implements Exception {
-  final String? reason;
-  SeedException([this.reason]);
-}
+
