+++ lib/database/exceptions/seed_exception.dart
@@ -1,0 +1,6 @@
+/// This exception is thrown when the database seeding
+/// is not as expected.
+class SeedException implements Exception {
+  final String? reason;
+  SeedException([this.reason]);
+}
