+++ lib/database/misc/companion_utils.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:nacht/data/data.dart';
+import 'package:nacht/database/database.dart';
 import 'package:nacht_sources/nacht_sources.dart' as sources;
 import 'package:drift/drift.dart';
