+++ lib/database/models/chapter.dart
