+++ lib/database/models/meta_entry.dart
